Update feature for backend done

+++ backend/src/main/java/com/example/inventory/repository/InventoryRepository.java
@@ -10,5 +10,5 @@ public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
 
     List<Inventory> findAllByIsDeletedFalse();
 
-    Optional<Inventory> findByIdAndIsDeletedFalse(Integer id);
+    Optional<Inventory> findByIdAndIsDeletedFalse(Long id);
 }
